fix : add delete and modify functions with images

+++ backend/controllers/stuff.js
@@ -1,5 +1,5 @@
 const Thing = require('../models/Thing'); // Import the Thing model
-
+const fs = require('fs');
 // Middleware for creating a thing
 exports.createThing = (req, res, next) => {
    const thingObject = JSON.parse(req.body.thing);
@@ -41,9 +41,22 @@ exports.modifyThing = (req, res, next) => {
 
 // Middleware for deleting a thing
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
+   Thing.findOne({ _id: req.params.id})
+       .then(thing => {
+           if (thing.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = thing.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Thing.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
 };
 
 // Middleware for getting one thing
